Tools: Move has_wname_prefix to tplgtool2.py
This patch moves the helper function has_wname_prefix() from
topo_vol_kcontrols.py to tplgtool2.py. It will be used by a
new script to find bytes controls.
