+++ tools/tplgtool2.py
@@ -1243,6 +1243,31 @@ class TplgGraph:
             pipelines.append((comps[0], [self._nodes_dict[name] for name in endpoints if name.startswith('PCM')]))
         return pipelines
 
+def has_wname_prefix(widget):
+    """Is the kcontrol name prefixed with the widget name? ("PGAxx" or "Dmicxx")
+    Check SOF_TKN_COMP_NO_WNAME_IN_KCONTROL_NAME"""
+
+    wname_elems = [
+        prv.elems
+        for prv in widget.priv
+        if prv.elems[0].token
+        == SofVendorToken.SOF_TKN_COMP_NO_WNAME_IN_KCONTROL_NAME.name
+    ]
+
+    if len(wname_elems) == 0:  # typically: topo v1
+        no_wname_prefix = 0
+    elif len(wname_elems) == 1:  # typically: topo v2
+        assert len(wname_elems[0]) == 1
+        no_wname_prefix = wname_elems[0][0].value
+    else:
+        assert False, f"Unexpected len of wname_elems={wname_elems}"
+
+    assert no_wname_prefix in (0, 1)
+
+    # Double-negation: "no_wname false" => prefix
+    return not no_wname_prefix
+
+
 if __name__ == "__main__":
     from pathlib import Path
